requirements.txt
flask==2.3.3
pyTelegramBotAPI==4.14.0
requests==2.31.0
python-dotenv==1.0.0 